% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaRFTrain-package.R
\docType{package}
\name{adaRFTrain-package}
\alias{adaRFTrain}
\alias{adaRFTrain-package}
\title{adaRFTrain: grid search for rf}
\description{
Use grid search with cross validation to search rf hyper-parameter and do rf and dt.
}
\keyword{internal}
