% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dt_train.R
\name{dt_train}
\alias{dt_train}
\title{dt_train}
\usage{
dt_train(
  df,
  formula = NULL,
  df_test = NULL,
  p = 0.8,
  params_ = list(method = "recursive.partition", split = c("deviance", "gini")),
  target = NULL
)
}
\arguments{
\item{df:}{total dataframe if df_test is not specified else train data frame}

\item{formula:}{the formula of the model, can be not speicified if target is specified}

\item{df_test:}{test data frame}

\item{p:}{the ratio of split, have to be specified if df_test is not, default: 0.8}

\item{params_:}{the parameters of decision tree, have to be a list with method and split:}

\item{method:}{character string giving the method to use. default: "recursive.partition"}

\item{split:}{Splitting criterion to use. default: c("deviance", "gini")}

\item{target:}{the target of the model, have to be specified if formula is not}
}
\description{
This is the function to train and perdict the decision tree

which returns a dataframe with test data and result
}
