% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_search.R
\docType{data}
\name{params}
\alias{params}
\title{grid_search_rf}
\format{
An object of class \code{list} of length 7.
}
\usage{
params
}
\arguments{
\item{df:}{total dataframe}

\item{params_range:}{a list of parameters to tune}

\item{target:}{the target of the model}

\item{n_fold:}{number of folds to use to split the train data, default: 3}

\item{scoring:}{scoring metric used to evaluate the best model, default: c("accuracy", "auc")}

\item{params_:}{the parameters of random forest, have to be a list with all parameters:}

\item{ntree:}{Number of trees to grow, default: 500}

\item{sampsize:}{Size(s) of sample to draw, default: total row of train dataset}

\item{nodesize:}{Minimum size of terminal nodes, default: NULL}

\item{maxnodes:}{Maximum number of terminal nodes trees in the forest can have, default: NULL}

\item{replace:}{Should sampling of cases be done with or without replacement?, default: TRUE}

\item{mtry:}{Number of variables randomly sampled as candidates at each split,default: floor of square of features}

\item{strata:}{A (factor) variable that is used for stratified sampling, default: target}
}
\description{
This is the function to search the hyper parameter of the random forest

which returns a total hyper parameter for random forest
}
\keyword{datasets}
