% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rf_train.R
\name{rf_train}
\alias{rf_train}
\title{dt_train}
\usage{
rf_train(
  df,
  formula = NULL,
  df_test = NULL,
  p = 0.8,
  params_ = list(ntree = 500, sampsize = NULL, nodesize = NULL, maxnodes = NULL,
    replace = TRUE, mtry = NULL, strata = NULL),
  target = NULL
)
}
\arguments{
\item{df:}{total dataframe if df_test is not specified else train data frame}

\item{formula:}{the formula of the model, can be not speicified if target is specified}

\item{df_test:}{test data frame}

\item{p:}{the ratio of split, have to be specified if df_test is not, default: 0.8}

\item{params_:}{the parameters of random forest, have to be a list with all parameters if specified:}

\item{ntree:}{Number of trees to grow, default: 500}

\item{sampsize:}{Size(s) of sample to draw, default: total row of train dataset}

\item{nodesize:}{Minimum size of terminal nodes, default: NULL}

\item{maxnodes:}{Maximum number of terminal nodes trees in the forest can have, default: NULL}

\item{replace:}{Should sampling of cases be done with or without replacement?, default: TRUE}

\item{mtry:}{Number of variables randomly sampled as candidates at each split, default: floor of square of features}

\item{strata:}{A (factor) variable that is used for stratified sampling, default: target}

\item{target:}{the target of the model, have to be specified if formula is not}
}
\description{
This is the function to train and predict the random forest

which returns a dataframe with test data and result
}
